Add files via upload
The value of customer retention should be a high priority for all businesses. One of the research of ‘Fred Reichard of Bain and Company’ found a 5% increase in customer retention can increase profits by over 25%. So developed a robust and accurate churn prediction model to retain customers and form strategies for reducing customer attrition rates.
Churn Prediction: Churn is the metric that measures the rate at which a business will lose customers.